CHAR: Switch from unsigned char to signed char
Once upon a time, as part of some suggestions by the Watcom Compiler, I had dared to switch the CHAR type default from signed to unsigned. I suppose modern C++ is trying to switch back to unsigned with the `char_t` type, but in terms of legacy C99 code trying to use an unsigned char* for strings leads to a lot of compiler warnings and surprising type mismatches and, and without tangible upside, is better left as a regular signed char.