Remove mm_forbid and disable psforbid - out-dated and ineffective threading model
If threading is to be implemented, it will need to be done using a message queue that submits player state changes to the backend. Modifying state vars inside locks is dangerous, hard to debug, and often results in big perf penalties due to the broad scope of time each lock needs to hold.