plugin: re-enable logging and output to both stdout and OutputDebugString
Can consider circling back to reduce scope of logging on release builds later, but for now the low volume of logging is such that it's likely not a concern ever. Back in 1995 it was actually rather important to avoid any extra logging as much as possible. A new era for Mikamp is upon us.

# Conflicts:
#	MikampPlugin/src/main.cpp